+++ CHANGELOG.md
@@ -8,6 +8,7 @@ TBD
 - Added max concurrent support for beer.conf under `MAX_CONCURRENT`
 - Fixed issue where using local systems with a Namespace not the Garden name would not be treated as a local
   Request for internal features
+- Added config option to import scheduled jobs file at startup
 
 # 3.27.2